Add factory for active stress models
