+++ Code/Source/solver/active_stress_uniform.cpp
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: Copyright (c) Stanford University, The Regents of the
+// University of California, and others. SPDX-License-Identifier: BSD-3-Clause
+
+#include "active_stress_uniform.h"
+
+REGISTER_ACTIVE_STRESS_MODEL("uniform", UniformActiveStress);
